Chart Data

+++ fishing-the-net/src/main/java/com/example/fishingthenet/home/HomeInfoDto.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class HomeInfoDto {
 
-    private float percentage;
+    private int percentage;
     private String mostfrequentCountry;
     private int numberOfFishingEmails;
 }
